# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QueuePilot API
  description: >
    Distributed job scheduler API. Submit one-shot or cron-recurring jobs (optionally
    depending on other jobs as a DAG), inspect their runs, and manage the worker fleet.
    See docs/ARCHITECTURE.md for design rationale and docs/VERIFICATION.md for real
    measured behavior. Generated by hand from internal/api/router.go and handlers.go —
    kept in sync manually, not code-generated.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local docker-compose deployment

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness/readiness check
      security: []
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /metrics:
    get:
      operationId: metrics
      summary: Prometheus metrics
      security: []
      responses:
        "200":
          description: Prometheus text exposition format
          content:
            text/plain:
              schema:
                type: string

  /v1/jobs:
    post:
      operationId: createJob
      summary: Create a job
      description: >
        Idempotent when idempotency_key is set: a retried request with the same key
        returns the original job (200) instead of creating a duplicate or erroring.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewJobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "200":
          description: Idempotent replay — returns the existing job for this idempotency_key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: >
            Validation failed: name is required / invalid cron_expr /
            max_attempts must be >= 1 / timeout_seconds must be >= 1 / invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: idempotency_key already used by a different job (race on first use)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      operationId: listJobs
      summary: List jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, archived]
        - name: limit
          in: query
          description: Default 50, capped at 200
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Jobs matching the filter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "400":
          description: invalid status / invalid limit / invalid offset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/jobs/{id}:
    get:
      operationId: getJob
      summary: Get a job by ID
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: The job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/jobs/{id}/pause:
    post:
      operationId: pauseJob
      summary: Pause a job (sets status=paused; the promoter stops scheduling new runs for it)
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/jobs/{id}/resume:
    post:
      operationId: resumeJob
      summary: Resume a paused job (sets status=active)
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/jobs/{id}/runs:
    get:
      operationId: listJobRuns
      summary: List runs for a job
      description: >
        Returns an empty array (not 404) if the job ID doesn't exist — this endpoint
        doesn't check job existence first, it just queries runs by job_id.
      parameters:
        - $ref: "#/components/parameters/JobId"
        - name: limit
          in: query
          description: Default 50, capped at 200
          schema:
            type: integer
      responses:
        "200":
          description: Runs for this job, most recent first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRun"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/runs/{id}:
    get:
      operationId: getRun
      summary: Get a run by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRun"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/workers:
    get:
      operationId: listWorkers
      summary: List known workers and their last heartbeat
      responses:
        "200":
          description: Workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256 JWT signed with the deployment's shared JWT_SECRET. This is a
        single-admin-credential model, not multi-tenant — every valid token has full
        access to every job (see docs/ARCHITECTURE.md "known limitations"). Mint one
        with internal/api/auth.go's MintToken(secret, subject, ttl) — there is no
        login endpoint.

  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Unauthorized:
      description: Missing, malformed, invalid, or expired bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: >
        Per-client-IP token bucket exhausted (default 20 req/s, burst 40 — see
        RATE_LIMIT_RPS / RATE_LIMIT_BURST)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string

    NewJobInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Also the handler registry key the worker looks up (e.g. "echo", "sleep", "http_call")
        payload:
          type: object
          additionalProperties: true
        cron_expr:
          type: string
          nullable: true
          description: Standard 5-field cron expression. Omit for a one-shot job.
          example: "*/5 * * * *"
        priority:
          type: integer
          format: int16
          default: 0
          description: Higher runs first
        max_attempts:
          type: integer
          format: int16
          default: 5
          minimum: 1
        timeout_seconds:
          type: integer
          format: int32
          default: 300
          minimum: 1
        idempotency_key:
          type: string
          nullable: true
        depends_on:
          type: array
          items:
            type: string
            format: uuid
          description: Job IDs that must have a succeeded run before this job's runs are promoted

    Job:
      allOf:
        - $ref: "#/components/schemas/NewJobInput"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [active, paused, archived]
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    JobRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, leased, running, succeeded, failed, dead]
        attempt:
          type: integer
          format: int16
        priority:
          type: integer
          format: int16
        scheduled_at:
          type: string
          format: date-time
        leased_by:
          type: string
          nullable: true
        leased_at:
          type: string
          format: date-time
          nullable: true
        lease_expires_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          nullable: true
          additionalProperties: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    Worker:
      type: object
      properties:
        id:
          type: string
        hostname:
          type: string
        status:
          type: string
          enum: [alive, dead]
        last_heartbeat:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time